Zoltan: Fixed build warnings; added comments to some.
The best way to wake up is to fix build warnings.  It's good therapy.

One of the build warnings relates to the correct printf specifier for
an int64_t.  %lld is for long long, not for int64_t.  inttypes.h
provides a macro to use.  Stack Overflow has reported some issues with
Windows not implementing these macros (i.e., not having a correct C99
implementation).  I omitted the suggested macrology to fix this, but
please note that it could be an issue.