Ifpack2: LocalFilter: apply: Added check for X aliasing Y.
I was having some troubles where enabling the Kokkos Refactor version
of Tpetra was causing the RILUK test to fail.  In order to check what
was going on, I added a test in LocalFilter for whether X and Y alias
one another (in an obvious way, that is, &X == &Y -- it's not a test
of all possible ways they can alias one another).  This test is a good
idea anyway, so I left it in place.  The apply() method should now
work whether or not &X == &Y.

Build/Test Cases Summary
Enabled Packages: Tpetra, Ifpack2
0) MPI_DEBUG => passed: passed=92,notpassed=0 (5.24 min)
1) SERIAL_RELEASE => passed: passed=68,notpassed=0 (4.38 min)
2) MPI_RELEASE => passed: passed=92,notpassed=0 (8.48 min)
Other local commits for this build/test group: cdf2415, 1ded3cf