Ifpack2: loosening test tolerance
On a few machines, the tolerance used in comparison of the norms of
the inverse of the diagonals in the ILU(k) tests is apparently too stringent.
Increasing it to 1e-7.  Also cleaned up printing in another test so that
if a failure occurs, it's clear which output goes with the error.

Build/Test Cases Summary
Enabled Packages: Ifpack2
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=25,notpassed=0 (16.95 min)
1) SERIAL_RELEASE => passed: passed=20,notpassed=0 (8.50 min)