fortran underscore fixes for bgq - From Dan Sunderland...
There were a lot of places where we were relying on platform or compiler
defines to set the underscore decoration requirements for fortran. Since
bgq uses gcc, but the underscaore requirements need to match vacpp this
is non-functional.  This changes those header files to use the macro
defined by bjam which correctly disambiguates the 2 versions of gcc.

(From Sierra)