Tpetra: MultiVector: Fixed view semantics version of operator=.
In the classic version of Tpetra::MultiVector, if the right-hand side
of an assignment has view semantics, it transmits view semantics to
the left-hand side of the assignment.  In this case in operator=, I
wasn't passing the right-hand side's Kokkos Node instance to the
left-hand side.  This showed up when I was adding a new feature to
Tpetra::Experimental::BlockMultiVector.  This commit fixes the
omission, and also simplifies the code.