Tpetra: BlockCrsMatrix: Fixed uncertainty about rel vs. abs offsets.
In the ongoing implementation of Tpetra::Experimental::BlockCrsMatrix,
there was some uncertainty about whether internal offsets were
relative to the row, or absolute over the whole matrix.  Mixing up
relative and absolute offsets could cause difficult-to-debug segfaults
and memory corruption.  In order to prevent this, I changed names of
methods to make this difference more explicit.  The method that
searches within a row for a column index's offset now takes the hint
as a relative offset, and returns a relative offset, for consistency
with CrsGraph.