Panzer: Fixing edge bases so they are always oriented in a consistent way
The way panzer does dirichlet boundary conditions (there is a reason for this)
is to sum them across processors and then scale them back down. For the
edge basis with the orientations one processor would set this as (say)
0.5 and the other (with opposite orientation) would use -0.5. Thus when
you summed acrossed processors you get 0.0 (which is clearly incorrect).
The solution (proposed) is to multiply by the orientation vector by the
dirichlet residual. This will preserve residuals with orientation 1.0, and
flip the sign on edges with orientation of -1.0. So they should be consistent
acrossed processors giving the correct summation.