Panzer: Adding a setOnetimeDirichletBeta method that uses recursion
This method recurses through a hierarchy of model evaluators (hiarchy
meaning nesting in this context) to set the beta parameter. This assumes
that if the ME is not a panzer specfic ME (Thyra or Epetra) then it
must be a delegator. The method then recurses on the underlying model
evaluator until it finds a panzer specific ME (and succeeds) or can't
find a delegator (and terminates with an error).