Tpetra:  Fix Tpetra::getData{NonConst}() methods for KokkosRefactor.
The persisting views returned by these methods could be the wrong length
due to padding of the underlying Kokkos view's.  This is because
persisitingView() uses capacity() to compute the ArrayRCP's size instead
of size(), which was done explicitly a while to ago to make sure the
ArrayRCP had the right size for views of rank > 1.  So to fix this I set
the size of the returned persiting view explicitly.  This fixes a number
of MueLu failures.