Stokhos:  Check for non-contiguous allocations in PCE view spec.
When Kokkos manages the PCE allocation, it is always contiguous.  But
the view can also take a third-party pointer, which may not have the
right allocation.  This happens, e.g., when MueLu creates its tentative
prolongator.  We know check for this, and throw errors when an operation
is done that doesn't work without contiguous allocations, and implements
a deep_copy() that does.  Added a unit test for this scenario.