Kokkos: Ease of use improvements
Added a central Kokkos_Core.hpp file which includes all the relevant
headers depending on which devices are configured.
This file also provides a Kokkos::initialize() and Kokkos::finalize()
which will call the corresponding functions of DefaultDeviceType and
its host_mirror_device_type. We need to discuss how to add the
variants of those functions which take arguments.

Made UnitTest being instantiated with DefaultDeviceType.

The ease of use stuff allows now this (with C++11):

int main() {
  Kokkos::initialize();

  Kokkos::parallel_for(10,[=] (int i) {
    printf("HelloWorld %i\n",i);
  });

  Kokkos::finalize();
}

Doing this with a functor allows running on GPUs as well, the functor would be:

struct helloworld {
  KOKKOS_INLINE_FUNCTION
  void operator() (int i) const {
    printf("HelloWorld %i\n",i);
  }
}