Tpetra: MV: Changed getLocalMV's ret type & deprec'd getLocalMVNonConst.
In the Kokkos refactor version of Tpetra,
MultiVector::getLocalMVNonConst was already deprecated.  This commit
adds that deprecation to the "classic" version of Tpetra.  It also
changes the return type of getLocalMV (in the classic version of
Tpetra; this was already done in the Kokkos refactor version) from
"const KokkosClassic::MultiVector&" to KokkosClassic::MultiVector.

This commit introduces build warnings to Ifpack2, due to use of the
deprecated method.  Subsequent commit(s) will remove use of the
deprecated method in Ifpack2.

I made this change in order to simplify completion of the Kokkos
refactor port of Tpetra.  There was never a need to have separate
getLocalMV and getLocalMVNonConst functions.  This is because

  1. the intent of each was to access the data, not to change the
     pointer (e.g., by reallocating), and
  2. KokkosClassic::MultiVector always implemented lightweight view
     semantics, so there was never a need to return it by reference.

The const method getLocalMV would have sufficed for both const and
nonconst access to the data, but only by returning the
KokkosClassic::MultiVector by value (shallow copy), not by reference.
This commit enforces this by deprecating getLocalMVNonConst, and
making getLocalMV return by value, not by reference.  The latter means
that getLocalMV can be used for both const and nonconst access to the
data.