Epetra: Added the third of five lessons.
I am translating the Tpetra lessons into Epetra, in preparation for
the VECPAR tutorial at the end of this month.  This commit adds the
third of the five lessons.

This lesson was derived from the power method example in Tpetra, which
in turn was derived from the classic power method example in Epetra.
The latter lives in epetra/example/petra_power_method (32-bit global
indices) and epetra/example/petra_power_method_LL (64-bit global
indices).  The lesson added in this commit subsumes both cases, so
both the petra_power_method and petra_power_method_LL directories may
be safely removed.