STK: Fix Trilinos build issues in Classic
Fix option names to depend on the package name which is now STKClassic instead
of just STK. This was an issue since automatic options in STKClassic have
changed due to the move.

Fixed an issue with the install where the generated configure file was not being
installed due to a name change.

(From Sierra)