Kokkos: Fix to support relocatable device code
When compiling with CUDA and using relocatable device code:
--relocatable-device-code true
the const buffer for functors can only be defined once.
Thus it needs to be declared as an external resource in the header file
and added to the impl.cu file in Kokkos core/Cuda. When not compiling
with relocatable device code this approach does not work, because this
external symbol will be invalid. Relocatable device code is necessary
for function pointers (e.g. if device functions are compiled in one
compilation unit and called from a global function in another compilation
unit). Even if we don't use those in Kokkos Core and application might
use them. I added a flag KOKKOS_CUDA_USE_RELOCATABLE_DEVICE_CODE which
has to be defined to add the "extern" qualifier. There is no CMake option
for this at the moment.