Tpetra: Added test for Bug 6170.
Thanks to Clark Dohrmann for contributing this test.  I modified it
slightly to avoid file I/O and to catch and print exception messages
on all processes.  The test must be run with exactly 8 MPI processes.

Interestingly enough, on my workstation (GNU/Linux, recent gcc and
OpenMPI), this test fails for the "classic" version of Tpetra, but
passes for the "Kokkos refactor" version of Tpetra.

NOTE: The default behavior of the Trilinos test framework is to
exclude tests that require more than 4 MPI processes.  Thus, this test
will NOT run by default.  To change this, set the CMake option
MPI_EXEC_MAX_NUMPROCS to 8 or more in your Trilinos configuration.