Tpetra::Experimental::BlockCrsMatrix: Finished DistObject methods.
This commit adds a test that shows that the "copy" part of
copyAndPermute works.  All the other DistObject methods needed for
Import and Export to work have been implemented, and Import runs.
However, we don't yet have tests in place that exercise the "permute"
part of copyAndPermute, or the combine modes.

This commit also adds description() and describe() methods to
BlockCrsMatrix.  If you call describe() with verbLevel =
Teuchos::VERB_EXTREME, it prints the entire matrix in a human-readable
way.  (The method only actually prints on Process 0 of the matrix's
communicator; all processes send their data to Process 0 for printing
in a memory-scalable way.)  This is useful for debugging, and also
answers a feature request by Travis.