Repurpose edge marking to work with all entities
This is mostly just a renaming, since the previous implementation
already managed a marking array that contained all entities.  Some
code to zero out this entire array had to be reworked to just zero out
the edges, since it will also manage node sharing in an upcoming commit.

(From Sierra)