Tpetra: Ported Zoltan2 RCB Epetra/Tpetra/Zoltan comparison to Tpetra.
The following test in Zoltan2:

Trilinos/packages/zoltan2/test/temp/multivectorTest.cpp

compares the performance of Epetra, Tpetra, and Zoltan(1) data
migration, for a use case resembling what recursive coordinate
bisection might do.  That test was crashing with the Kokkos refactor
version of Tpetra.  Thus, I ported just the Tpetra parts of the test
to Tpetra.  I then found out that the Zoltan2 test was using
Tpetra::MultiVector incorrectly.

In a debug build (with Teuchos_ENABLE_DEBUG ON), Teuchos::RCP catches
a dangling reference.  This is because the code calls
Tpetra::MultiVector::getData(j), which returns an
Teuchos::ArrayRCP<const Scalar>, but then keeps a nonowning view (as a
Teuchos::ArrayView<const Scalar>) of that ArrayRCP.  This is illegal,
because the ArrayRCP has fallen out of scope.

The only way one would see erroneous behavior in this case is if Node
= KokkosClassic::ThrustGPUNode.  In that case, the returned ArrayRCP
is a host allocation ("generalized host view of GPU device data") and
would actually be freed when the ArrayRCP's reference count goes to
zero.  This case could also break the Kokkos refactor Nodes that have
separate memory spaces (currently only the CUDA wrapper Node).

I fixed this issue in the Tpetra port of the RCB test, and it runs
without crashing or throwing an exception.  The fix actually makes the
code simpler and easier to understand.  Thus, I would recommend
porting the fix to Zoltan2.  I will do it myself as soon as I get the
chance.

I also took the chance to clean up some of the nonidiomatic use of
Teuchos::Time and other Teuchos classes.  This clean-up should also be
ported back to the Zoltan2 test.