Tpetra: MultiVector (Kokkos ref): Fixed deep_copy & added error checks
This commit only affects the implementation of the Kokkos refactor
version of MultiVector.

I fixed some semantic issues with deep_copy for MultiVector.  First,
it now no longer requires UVM to work with CUDA, because it does
sequential fill of whichVectors on the host and uses Kokkos::DualView
to push them to the device (if necessary).  Second, it now correctly
syncs between host and device for the destination MultiVector.

I also added error checks to MultiVector that are more appropriate for
the Kokkos refactor version.