Updated to use GL_USER if set for gitolite (PHI Kanban #3419)
Here I updated this simple bash script to use GL_USER set in the env
if it is non-empty instead of USER.  The env var GL_USER is set by
gitolite on a push.  This maintains the old behavior when GL_USER is
undefined and will use the push userid when gitolite is being used.

From a Google search, it looks like GL_USER is pretty much only used
by Gitolite so I don't think there is much of a chance of accidentally
picking a var called GL_USER unless gitolite is being used so I think
this is safe.

Yea, now the push hooks work for gitolite too!