Ifpack2: addititional test for AdditiveSchwarz
This test exercises Amesos2 SuperLU as a subdomain solve to Additive
Schwarz.  The solution is compared to a solution from Additive Schwarz
using a dense direct solve as a subdomain solve.  They should be the same.

Also made minor change to Amesos2solver -- named a sublist "Amesos2",
which is required by Amesos2.