Kokkos: Add compatibility checks to Kokkos::deep_copy implementation to catch errors.
An outstanding problem to be addressed is a call to deep_copy with
fully compatible subviews of rank > 1 that have strided subranges
smaller than the dimensions will copy full range of contiguous
memory spanned by the subviews.  This copies too much.  Challenge
is to keep the high-performing deep_copy for all compatible views
that do not have this sub-strided characteristic.