PyTrilinos: Fixed warnings from LOCA.AnasaziOperator wrapper
There was a simple label() method that returned a std::string& that
was causing a warning because the class was wrapped as a director, and
you shouldn't return references from a director method.  I implemented
a simple redefinition of the method to return a copy, and that fixed
the warning.