decomp: Add workaround for muzia decomp issues
On some machines (like muzia) the launch script doesn't work on the
login node; it must be submitted to the queue before it can be run. This
actually worked on muzia until a recent OS upgrade, but according to the
HPC admins it was luck that it worked, and not guaranteed to work. :(

This caused failures when running decomp on muzia, since it is simply a
shell script that uses launch (if it is in PATH) to run various seacas
apps.

This commit presents a workaround for this failure. What is does it to
test the launch script with a simple command (true) to see if it works
on the node we are on. If it doesn't it will not use it to run the other
apps. The command 'launch -n 1 true' works on all the SRN HPC machines,
and on the SRN desktop machines, with the exception of muzia and bgq,
and on both of those nem_slice works on the login node.

This result of this change is that decomp will attempt to run the
various seacas apps directly on nodes where launch doesn't work, and it
will still attempt to launch them on nodes where launch does work.
So, this change should be no worse than the current behavior on most
machines, and should allow the nightly process to work on muzia and bgq.
Clearly if approved this change would need to be made in the branches
as well.

Change-Id: I6a2a76e29f0777cd5f8c67850b5ffecc127658f9

(From Sierra)