Tpetra: CrsGraph: reindexColumns() now works in Kokkos refactor
This commit mainly only affect the Kokkos refactor version of Tpetra,
despite the minor change to the "classic" version of CrsGraph.

The new CrsGraph::reindexColumns() method now passes its test, with
the Kokkos refactor version of CrsGraph, in a release build of Tpetra.

If Teuchos_ENABLE_DEBUG is ON, however, this test fails in
CrsGraph::clone.  A possible problem is that the Kokkos refactor
version of CrsGraph creates and returns Teuchos::ArrayView of a raw
pointer, and the debug-mode memory tracking in ArrayView doesn't like
this and throws an exception, thinking that it's a dangling memory
reference.  Eric Phipps, Ross Bartlett, and I are working on resolving
this.

Build/Test Cases Summary
Enabled Packages: KokkosCore, KokkosCompat, KokkosContainers, KokkosLinAlg, KokkosMpiComm, KokkosAlgorithms, Tpetra
0) MPI_DEBUG => passed: passed=75,notpassed=0 (5.44 min)
1) SERIAL_RELEASE => passed: passed=54,notpassed=0 (4.96 min)
2) MPI_RELEASE_KOKKOS_REFACTOR => passed: passed=77,notpassed=0 (9.85 min)
3) SERIAL_RELEASE_KOKKOS_REFACTOR => passed: passed=64,notpassed=0 (9.76 min)
Other local commits for this build/test group: f02817c