Tpetra: CrsMatrix: Replaced LocalMatOps use for local Gauss-Seidel.
This commit only affects the Kokkos refactor version of Tpetra.

The local part of Gauss-Seidel now no longer uses the LocalMatOps
instance in CrsMatrix.  This is one step in getting rid of the
LocalMatOps instance.  Once it's gone from CrsMatrix and CrsGraph, we
will be able to deprecate use of KokkosClassic in Tpetra.