Panzer: Fixes for the incorrectly sized AD type
Thanks to EP for finding this. The issue was that the size for the field
type was based on a single (vector) PDE. This was not the case in his
mixed formulation so of course it failed. The fix was to simply figure
out the size of the AD types being passed in as dependencies.