Disabling 3 Intrepid tests that are timing out after 3 minutes (Trilinos #6224)
The thee Intrepid tests

    Intrepid_test_Discretization_Basis_HGRAD_TRI_Cn_FEM_Test_02_MPI_1 (Timeout)
    Intrepid_test_Discretization_Basis_HDIV_TET_In_FEM_Test_02_MPI_1 (Timeout)
    Intrepid_test_Discretization_Basis_HGRAD_TET_Cn_FEM_Test_02_MPI_1 (Timeout)

timeout in the default PT MPI_DEBUG build.  This is stopping my pushes so I
must disable them for now.  See Trilinos #6224 for more details.

These will stay disabled in this script until such time that these tests can
be made to finish and pass in less than 3 minutes on this machine.